+++ src/detection/gpu/gpu_obsd.c
@@ -42,13 +42,13 @@ const char* ffDetectGPUImpl(FF_MAYBE_UNUSED const FFGPUOptions* options, FFlist*
             for (uint8_t func = 0; func <= maxfuncs; func++)
             {
                 uint32_t pciid, pciclass;
-                if (pciReadConf(pcifd, bus, dev, func, PCI_ID_REG, &pciid) != 0)
+                if (pciReadConf(pcifd, (uint8_t) bus, dev, func, PCI_ID_REG, &pciid) != 0)
                     continue;
 
                 if (PCI_VENDOR(pciid) == PCI_VENDOR_INVALID || PCI_VENDOR(pciid) == 0)
                     continue;
 
-                if (pciReadConf(pcifd, bus, dev, func, PCI_CLASS_REG, &pciclass) != 0)
+                if (pciReadConf(pcifd, (uint8_t) bus, dev, func, PCI_CLASS_REG, &pciclass) != 0)
                     continue;
 
                 if (func == 0)
